+++ sbuild/__init__.py
@@ -1,0 +1,1 @@
+from sbuild.tools import compiler, linker