+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Dates are in yyyy-mm-dd format.
 - Added base node class and basic functionality.
 - Added linker class, and 2 default linkers: GCC and Clang.
 - Disables logging when python is run with -O. This can provide some speed improvements.
-- Added Source, Header and Object nodes.
+- Added Source and Object nodes.
 
 ## SBuild v0.1.0 (2019-02-16)
 - Initial version, basic compiler functionality.